reduce reliance on Qt with callbacks
