+++ src/app/CMakeLists.txt
@@ -8,6 +8,6 @@ add_executable(main
 
 target_link_libraries(main 
     Qt6::Widgets
-    algorithm_lib
+    service_lib
     ui
 )
