+++ src/app/ui/mainwindow.ui
@@ -6,7 +6,7 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>900</width>
+    <width>920</width>
     <height>700</height>
    </rect>
   </property>
